requirements.txt
flask==2.3.3
flask-cors==4.0.0
tensorflow==2.19.0
pillow==10.0.0
numpy==1.24.3
matplotlib==3.7.2
requests==2.31.0
scikit-learn==1.3.0
